STEP 15. Move out batch opd and dragdrop
