+++ altitude/static/js/frontend-app.js
@@ -1,12 +1,8 @@
 import { Const } from "./constants.js"
 import { Folder } from "./models/folder.js"
 import { showErrorSnackBar } from "./common/snackbar.js"
-import {
-    dragged,
-    dragMoveListener,
-    setFixedPositionWhileDragging,
-} from "./common/dragon-drop.js"
 import { createAssetActions } from "./assets/asset-actions.js"
+import { bindAppDragDrop } from "./dragdrop/index.js"
 import { hydrateAppFragments } from "./fragments/index.js"
 import { handleViewSettingChanged } from "./fragments/search-results.js"
 import {
@@ -49,9 +45,7 @@ export class FrontendApp {
         this.initializeStores()
         this.registerEventListeners()
         this.Alpine.start()
-        this.bindBatchOpsDragDrop()
-        this.bindPeopleDragDrop()
-        this.bindFolderDragDrop()
+        bindAppDragDrop(this)
         this.hydrateFragments(document)
 
         return this
@@ -231,186 +225,6 @@ export class FrontendApp {
         this.hydrateFragments(event.detail.target)
     }
 
-    bindBatchOpsDragDrop() {
-        const dragHandleEl = document.querySelector("#batchOps button.drag-drop")
-        if (!dragHandleEl) {
-            return
-        }
-
-        interact("#batchOps button.drag-drop").draggable({
-            inertia: true,
-            autoScroll: true,
-
-            listeners: {
-                move: dragMoveListener,
-                start: () => {
-                    const selectedAssetsStore = this.Alpine.store(
-                        Const.state.selectedAssets,
-                    )
-
-                    selectedAssetsStore.items.forEach((asset) => {
-                        asset.drag()
-                    })
-                },
-                end: (event) => {
-                    const selectedAssetsStore = this.Alpine.store(
-                        Const.state.selectedAssets,
-                    )
-
-                    selectedAssetsStore.items.forEach((asset) => {
-                        asset.drop()
-                    })
-
-                    dragged(event)
-                },
-            },
-        })
-    }
-
-    bindPeopleDragDrop() {
-        interact("#people .drag-drop, #person .drag-drop").draggable({
-            inertia: true,
-            autoScroll: true,
-
-            listeners: {
-                move: dragMoveListener,
-                start: setFixedPositionWhileDragging,
-                end: dragged,
-            },
-        })
-
-        interact("#people .dropzone, #person.dropzone").dropzone({
-            accept: "#person .drag-drop, #people .drag-drop",
-            overlap: 0.75,
-
-            ondropactivate: (event) => {
-                event.target.classList.add("drop-active")
-            },
-
-            ondragenter: (event) => {
-                const draggableElement = event.relatedTarget
-                const dropzoneElement = event.target
-
-                dropzoneElement.classList.add("drop-target")
-                draggableElement.classList.add("can-drop")
-            },
-
-            ondragleave: (event) => {
-                event.target.classList.remove("drop-target")
-                event.relatedTarget.classList.remove("can-drop")
-            },
-
-            ondrop: (event) => {
-                const draggableElement = event.relatedTarget
-                const dropzoneElement = event.target
-
-                dropzoneElement.classList.remove("drop-active")
-                dropzoneElement.classList.remove("drop-target")
-                draggableElement.classList.remove("can-drop")
-
-                const mergeSourceId = draggableElement.getAttribute(
-                    Const.attributes.personId,
-                )
-                const mergeDestId = dropzoneElement.getAttribute(
-                    Const.attributes.personId,
-                )
-
-                console.debug(`Merging ${mergeSourceId} into ${mergeDestId}`)
-                this.dispatch(Const.events.confirmPersonMerge, {
-                    mergeSourceId,
-                    mergeDestId,
-                })
-            },
-
-            ondropdeactivate: (event) => {
-                event.target.classList.remove("drop-active")
-                event.target.classList.remove("drop-target")
-            },
-        })
-    }
-
-    bindFolderDragDrop() {
-        interact("#rootFolderList .drag-drop").draggable({
-            inertia: true,
-            autoScroll: true,
-
-            listeners: {
-                move: dragMoveListener,
-                end: dragged,
-            },
-        })
-
-        interact("#rootFolderList .dropzone").dropzone({
-            accept: "#rootFolderList .drag-drop, #assets .drag-drop, #batchOps .drag-drop",
-            overlap: 0.2,
-
-            ondropactivate: (event) => {
-                event.target.classList.add("drop-active")
-            },
-
-            ondragenter: (event) => {
-                const draggableElement = event.relatedTarget
-                const dropzoneElement = event.target
-
-                dropzoneElement.classList.add("drop-target")
-                draggableElement.classList.add("can-drop")
-            },
-
-            ondragleave: (event) => {
-                event.target.classList.remove("drop-target")
-                event.relatedTarget.classList.remove("can-drop")
-            },
-
-            ondrop: (event) => {
-                const draggableElement = event.relatedTarget
-                const dropzoneElement = event.target
-                const movedFolderId = draggableElement.getAttribute(
-                    Const.attributes.folderId,
-                )
-                const movedAssetId = draggableElement.getAttribute(
-                    Const.attributes.assetId,
-                )
-                const isBatchMover = draggableElement.parentNode.classList.contains(
-                    "batch-mover",
-                )
-                const newParentId = dropzoneElement.getAttribute(
-                    Const.attributes.folderId,
-                )
-
-                dropzoneElement.classList.remove("drop-active")
-                dropzoneElement.classList.remove("drop-target")
-                draggableElement.classList.remove("can-drop")
-
-                if (movedFolderId) {
-                    console.debug(`Moved folder ${movedFolderId} to ${newParentId}`)
-                    this.dispatch(Const.events.folderMoved, {
-                        movedFolderId,
-                        newParentId,
-                    })
-                }
-
-                if (movedAssetId) {
-                    console.debug(`Moved asset ${movedAssetId} to ${newParentId}`)
-                    this.dispatch(Const.events.assetMoved, {
-                        assetId: movedAssetId,
-                        folderId: newParentId,
-                    })
-                }
-
-                if (isBatchMover) {
-                    console.debug(`Batch moving assets to folder ${newParentId}`)
-                    this.dispatch(Const.events.batchAssetsMoved, {
-                        folderId: newParentId,
-                    })
-                }
-            },
-
-            ondropdeactivate: (event) => {
-                event.target.classList.remove("drop-active")
-                event.target.classList.remove("drop-target")
-            },
-        })
-    }
 
     hydrateFragments(root) {
         hydrateAppFragments({ root, app: this })
